requirements.txt
altair==5.5.0
APScheduler==3.11.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
debugpy==1.8.13
feedparser==6.0.11
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
multitasking==0.0.11
narwhals==1.29.1
nltk==3.9.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
protobuf==5.29.3
pyarrow==19.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
setuptools==75.8.2
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.6
streamlit==1.43.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
yfinance==0.2.54
